requirements.txt
google-generativeai 
langchain-google-genai
crewai[tools]
crewai
langchain
pypdf2
python-dotenv